updating annotations and refactoring
