+++ sbmlutils/dfba/diauxic_growth/model_factory.py
@@ -89,7 +89,7 @@ from sbmlutils import fbc
 
 from sbmlutils import factory as mc
 from sbmlutils.report import sbmlreport
-from sbmlutils import annotation
+from sbmlutils.annotation import annotator
 
 from sbmlutils.dfba import builder
 from sbmlutils.dfba import utils
@@ -188,7 +188,6 @@ UNIT_FLUX = 'mmol_per_h'
 UNIT_FLUX_PER_G = 'mmol_per_h'  # !!! FIXME (unit scaling between models)
 
 
-########################################################################################################
 def fba_model(sbml_file, directory, annotations=None):
     """ Create FBA submodel.
 
@@ -258,7 +257,7 @@ def fba_model(sbml_file, directory, annotations=None):
 
     # write SBML file
     if annotations:
-        annotation.annotate_sbml_doc(doc, annotations)
+        annotator.annotate_sbml_doc(doc, annotations)
     sbmlio.write_sbml(doc, filepath=pjoin(directory, sbml_file), validate=True)
 
     return doc
@@ -347,7 +346,7 @@ def bounds_model(sbml_file, directory, doc_fba=None, annotations=None):
     mc.create_objects(model, objects)
 
     if annotations:
-        annotation.annotate_sbml_doc(doc, annotations)
+        annotator.annotate_sbml_doc(doc, annotations)
     sbmlio.write_sbml(doc, filepath=pjoin(directory, sbml_file), validate=True)
 
 
@@ -384,7 +383,7 @@ def update_model(sbml_file, directory, doc_fba=None, annotations=None):
 
     # write SBML file
     if annotations:
-        annotation.annotate_sbml_doc(doc, annotations)
+        annotator.annotate_sbml_doc(doc, annotations)
     sbmlio.write_sbml(doc, filepath=pjoin(directory, sbml_file), validate=True)
 
 
@@ -469,7 +468,7 @@ def top_model(sbml_file, directory, emds, doc_fba=None, annotations=None):
 
     # write SBML file
     if annotations:
-        annotation.annotate_sbml_doc(doc, annotations)
+        annotator.annotate_sbml_doc(doc, annotations)
     sbmlio.write_sbml(doc, filepath=os.path.join(directory, sbml_file), validate=True)
 
     # change back into working dir
@@ -484,7 +483,7 @@ def create_model(output_dir):
     directory = utils.versioned_directory(output_dir, version=settings.VERSION)
 
     f_annotations = os.path.join(os.path.dirname(os.path.abspath(__file__)), settings.ANNOTATIONS_LOCATION)
-    annotations = annotation.ModelAnnotator.annotations_from_file(f_annotations)
+    annotations = annotator.ModelAnnotator.annotations_from_file(f_annotations)
 
     # create sbml
     doc_fba = fba_model(settings.FBA_LOCATION, directory, annotations=annotations)
