+++ sbmlutils/dfba/toy_atp/model_factory.py
@@ -16,7 +16,7 @@ from sbmlutils import comp
 from sbmlutils import fbc
 from sbmlutils import sbmlio
 from sbmlutils import factory as mc
-from sbmlutils import annotation
+from sbmlutils.annotation import annotator
 from sbmlutils.report import sbmlreport
 
 
@@ -160,7 +160,7 @@ def fba_model(sbml_file, directory, annotations=None):
     fbc.create_objective(model_fbc, oid="RATP_maximize", otype="maximize", fluxObjectives={"R3": 1.0}, active=True)
 
     if annotations:
-        annotation.annotate_sbml_doc(doc, annotations)
+        annotator.annotate_sbml_doc(doc, annotations)
 
     # write SBML
     sbmlio.write_sbml(doc, filepath=os.path.join(directory, sbml_file), validate=True)
@@ -203,7 +203,7 @@ def bounds_model(sbml_file, directory, doc_fba, annotations=None):
 
     # annotations
     if annotations:
-        annotation.annotate_sbml_doc(doc, annotations)
+        annotator.annotate_sbml_doc(doc, annotations)
 
     sbmlio.write_sbml(doc, filepath=os.path.join(directory, sbml_file), validate=True)
 
@@ -240,7 +240,7 @@ def update_model(sbml_file, directory, doc_fba=None, annotations=None):
 
     # write SBML file
     if annotations:
-        annotation.annotate_sbml_doc(doc, annotations)
+        annotator.annotate_sbml_doc(doc, annotations)
     sbmlio.write_sbml(doc, filepath=os.path.join(directory, sbml_file), validate=True)
 
 
@@ -322,7 +322,7 @@ def top_model(sbml_file, directory, emds, doc_fba, annotations=None):
 
     # write SBML file
     if annotations:
-        annotation.annotate_sbml_doc(doc, annotations)
+        annotator.annotate_sbml_doc(doc, annotations)
     sbmlio.write_sbml(doc, filepath=os.path.join(directory, sbml_file), validate=True)
 
     # change back the working dir
@@ -337,7 +337,7 @@ def create_model(output_dir):
     :return directory in which model files exist.
     """
     f_annotations = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'annotations.xlsx')
-    annotations = annotation.ModelAnnotator.annotations_from_file(f_annotations)
+    annotations = annotator.ModelAnnotator.annotations_from_file(f_annotations)
 
     directory = utils.versioned_directory(output_dir, version=settings.VERSION)
 
@@ -381,6 +381,5 @@ def create_model(output_dir):
     return directory
 
 
-########################################################################################################################
 if __name__ == "__main__":
     create_model(output_dir=settings.OUT_DIR)
