+++ sbmlutils/examples/models/annotation/results/annotation_example_8.xml
@@ -46,10 +46,10 @@
             </rdf:Bag>
           </dcterms:creator>
           <dcterms:created rdf:parseType="Resource">
-            <dcterms:W3CDTF>2019-06-03T23:57:29Z</dcterms:W3CDTF>
+            <dcterms:W3CDTF>2019-06-04T12:36:42Z</dcterms:W3CDTF>
           </dcterms:created>
           <dcterms:modified rdf:parseType="Resource">
-            <dcterms:W3CDTF>2019-06-03T23:57:29Z</dcterms:W3CDTF>
+            <dcterms:W3CDTF>2019-06-04T12:36:42Z</dcterms:W3CDTF>
           </dcterms:modified>
         </rdf:Description>
       </rdf:RDF>
@@ -100,7 +100,7 @@
             <rdf:Description rdf:about="#meta_ext">
               <bqbiol:is>
                 <rdf:Bag>
-                  <rdf:li rdf:resource="https://identifiers.org/bto/BTO:0000089"/>
+                  <rdf:li rdf:resource="bto/BTO:0000089"/>
                 </rdf:Bag>
               </bqbiol:is>
             </rdf:Description>
@@ -113,7 +113,8 @@
             <rdf:Description rdf:about="#meta_cyto">
               <bqbiol:is>
                 <rdf:Bag>
-                  <rdf:li rdf:resource="https://identifiers.org/go/GO:0005829"/>
+                  <rdf:li rdf:resource="go/GO:0005829"/>
+                  <rdf:li rdf:resource="https://en.wikipedia.org/wiki/Cytosol"/>
                 </rdf:Bag>
               </bqbiol:is>
             </rdf:Description>
@@ -126,7 +127,7 @@
             <rdf:Description rdf:about="#meta_pm">
               <bqbiol:is>
                 <rdf:Bag>
-                  <rdf:li rdf:resource="https://identifiers.org/go/GO:0005886"/>
+                  <rdf:li rdf:resource="go/GO:0005886"/>
                 </rdf:Bag>
               </bqbiol:is>
             </rdf:Description>
@@ -141,9 +142,9 @@
             <rdf:Description rdf:about="#meta_e__gal">
               <bqbiol:is>
                 <rdf:Bag>
-                  <rdf:li rdf:resource="https://identifiers.org/bigg.metabolite/gal"/>
-                  <rdf:li rdf:resource="https://identifiers.org/chebi/CHEBI:28061"/>
-                  <rdf:li rdf:resource="https://identifiers.org/vmhmetabolite/gal"/>
+                  <rdf:li rdf:resource="bigg.metabolite/gal"/>
+                  <rdf:li rdf:resource="chebi/CHEBI:28061"/>
+                  <rdf:li rdf:resource="vmhmetabolite/gal"/>
                 </rdf:Bag>
               </bqbiol:is>
             </rdf:Description>
@@ -191,7 +192,18 @@
       </assignmentRule>
     </listOfRules>
     <listOfReactions>
-      <reaction id="e__GLUT2_GAL" name="galactose transport [e__]" reversible="true" fast="false" compartment="pm">
+      <reaction metaid="meta_e__GLUT2_GAL" id="e__GLUT2_GAL" name="galactose transport [e__]" reversible="true" fast="false" compartment="pm">
+        <annotation>
+          <rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#" xmlns:dcterms="http://purl.org/dc/terms/" xmlns:vCard="http://www.w3.org/2001/vcard-rdf/3.0#" xmlns:vCard4="http://www.w3.org/2006/vcard/ns#" xmlns:bqbiol="http://biomodels.net/biology-qualifiers/" xmlns:bqmodel="http://biomodels.net/model-qualifiers/">
+            <rdf:Description rdf:about="#meta_e__GLUT2_GAL">
+              <bqbiol:is>
+                <rdf:Bag>
+                  <rdf:li rdf:resource="sbo/SBO:0000284"/>
+                </rdf:Bag>
+              </bqbiol:is>
+            </rdf:Description>
+          </rdf:RDF>
+        </annotation>
         <listOfReactants>
           <speciesReference species="e__gal" stoichiometry="1" constant="true"/>
         </listOfReactants>
