+++ sbmlutils/examples/models/assignment/model.py
@@ -3,7 +3,6 @@
 Simple assignment test case.
 """
 from sbmlutils.units import *
-from sbmlutils.sbo import *
 from sbmlutils.factory import *
 
 from sbmlutils.modelcreator import templates
