+++ sbmlutils/examples/models/dallaman/model.py
@@ -6,7 +6,6 @@ DallaMan2006
 # TODO: T2DM simulations (in current version not working)
 
 from sbmlutils.units import *
-from sbmlutils.sbo import *
 from sbmlutils.factory import *
 from sbmlutils.modelcreator import templates
