+++ sbmlutils/examples/models/demo/model.py
@@ -3,7 +3,6 @@
 Demo kinetic network.
 """
 from sbmlutils.units import *
-from sbmlutils.sbo import *
 from sbmlutils.factory import *
 from sbmlutils.modelcreator import templates
