+++ sbmlutils/examples/models/distrib/distrib_ex1.py
@@ -4,7 +4,6 @@ Distrib example.
 """
 
 from sbmlutils.units import *
-from sbmlutils.sbo import *
 from sbmlutils.factory import *
 from sbmlutils.modelcreator import templates
