+++ sbmlutils/examples/models/fbc/fbc_ex1.py
@@ -4,7 +4,7 @@ FBA example with exchange reactions.
 """
 
 from sbmlutils.units import *
-from sbmlutils.sbo import *
+from sbmlutils.annotation.sbo import *
 from sbmlutils.factory import *
 from sbmlutils.fbc import Objective
 from sbmlutils.modelcreator import templates
