+++ sbmlutils/utils.py
@@ -5,6 +5,38 @@ import warnings
 import time
 import libsbml
 import functools
+import uuid
+
+
+
+def create_metaid(sbase):
+    """ Creates a unique meta id.
+
+    Meta ids are required to store annotation elements.
+    """
+    hash_id = _create_hash_id(sbase)
+    return 'meta_{}'.format(hash_id)
+
+
+def _create_hash_id(sbase):
+    """ Creates unique hash id for sbase in model.
+
+    :param sbase:
+    :return:
+    """
+    # FIXME: This must be reproducible, so models don't change on recreation
+    if sbase and hasattr(sbase, 'getId') and sbase.isSetId():
+        hash_id = sbase.getId()
+    else:
+        hash_id = uuid.uuid4().hex
+
+    # the special case of the assignment rules (getId() returns getVariable())
+    if isinstance(sbase, libsbml.AssignmentRule) or isinstance(sbase, libsbml.RateRule):
+        hash_id = uuid.uuid4().hex
+    return hash_id
+
+
+
 
 
 def timeit(f):
